docs/Architecting Agentic AI Solutions/To be released/workshop 2/carbon-aware-agent-workshop-local/carbon-aware-agent-workshop/requirements.txt
python-dotenv
langchain
langchain-openai
langchain-community
langgraph
python-dateutil
